requirements.txt
aiohttp>=3.9.0
aioredis>=2.0.1
asyncio>=3.4.3
ccxt>=4.1.0
confluent-kafka>=2.3.0
cryptofeed>=2.3.2
fastapi>=0.104.0
grpcio>=1.59.0
grpcio-tools>=1.59.0
numpy>=1.24.0
pandas>=2.0.0
prometheus-client>=0.18.0
pydantic>=2.4.0
python-dotenv>=1.0.0
pyyaml>=6.0
redis>=5.0.0
rx>=3.2.0
uvicorn>=0.24.0
web3>=6.11.0
websockets>=12.0
scipy>=1.10.0
scikit-learn>=1.3.0
pyod>=1.1.0
ruptures>=1.1.8
hmmlearn>=0.3.0
stumpy>=1.12.0
adtk>=0.6.2
ta>=0.10.2
pandas-ta>=0.3.14b
finta>=1.3
vectorbt>=0.24.0